#13 abstract photo logic into a parent class, so it can be shared by both add and edit
